+++ gluoncv/utils/version.py
@@ -25,32 +25,3 @@ def check_version(min_version, warning_only=False):
             warnings.warn(msg)
         else:
             raise AssertionError(msg)
-
-
-def _require_mxnet_version(mx_version, max_mx_version='2.0.0'):
-    try:
-        import mxnet as mx
-        from distutils.version import LooseVersion
-        if LooseVersion(mx.__version__) < LooseVersion(mx_version) or \
-            LooseVersion(mx.__version__) >= LooseVersion(max_mx_version):
-            version_str = '>={},<{}'.format(mx_version, max_mx_version)
-            msg = (
-                "Legacy mxnet=={0} detected, some modules may not work properly. "
-                "mxnet{1} is required. You can use pip to upgrade mxnet "
-                "`pip install -U 'mxnet{1}'` "
-                "or `pip install -U 'mxnet-cu100{1}'`\
-                ").format(mx.__version__, version_str)
-            raise RuntimeError(msg)
-    except ImportError:
-        raise ImportError(
-            "Unable to import dependency mxnet. "
-            "A quick tip is to install via "
-            "`pip install 'mxnet-cu100<{}'`. "
-            "please refer to https://gluon-cv.mxnet.io/#installation for details.".format(
-                max_mx_version))
-
-def _deprecate_python2():
-    if sys.version_info[0] < 3:
-        msg = 'Python2 has reached the end of its life on January 1st, 2020. ' + \
-            'GluonCV has now dropped support for Python 2.'
-        raise DeprecationWarning(msg)
